% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\name{theme_publication}
\alias{theme_publication}
\title{Publication-ready ggplot2 theme}
\usage{
theme_publication(base_family = "Times New Roman", base_size = 10)
}
\arguments{
\item{base_family}{The font family to use. Default is "Times New Roman".}

\item{base_size}{The base font size in pt. Default is 10.}
}
\value{
A ggplot2 theme object.
}
\description{
An optimized ggplot2 theme that uses Times New Roman by default and
adjusts spacing and font sizes for document export to MS Word.
}
